requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
black==23.12.0
click==8.1.7
dj-config-url==0.1.1
dj-database-url==2.1.0
Django==5.0
djangorestframework==3.14.0
exceptiongroup==1.2.0
frozenlist==1.4.1
idna==3.6
iniconfig==2.0.0
multidict==6.0.4
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.1.0
pluggy==1.3.0
pytest==7.4.3
pytest-django==4.7.0
pytz==2023.3.post1
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.9.0
yarl==1.9.4
gunicorn>=21.2,<22.0
whitenoise[brotli]>=6.0,<7.0

# Uncomment these lines to use a Postgres database. Both are needed, since in production
# (which uses Linux) we want to install from source, so that security updates from the
# underlying Heroku stack image are picked up automatically, thanks to dynamic linking.
# On other platforms/in development, the precompiled binary package is used instead, to
# speed up installation and avoid errors from missing libraries/headers.
psycopg[c]; sys_platform == "linux"
psycopg[binary]; sys_platform != "linux"
